Trung

+++ BankManagement.WebAPI/Entities/Deal.cs
@@ -8,7 +8,7 @@ namespace BankManagement.WebAPI.Entities
     public class Deal
     {
         public int DealId { get; set; }
-        public int Money { get; set; }
+        public float Money { get; set; }
         public DateTime Date { get; set; }
         public int CustomerIdSend { get; set; }
         public int CustomerIdRevice { get; set; }
